requirements.txt
streamlit
supabase
pandas
numpy
plotly
talib
pandas_ta
python-dotenv
bcrypt
ta
